Finish implementation of projectile.
The projectile can be fired, and it collides with objects correctly. Some notes; first, the bullets cluster together if the player fires in the direction they are moving. The current workaround is to make projectiles not check for collision with each other; another solution may be preferable. Second, several changes were made to the player controller script in order to instantiate the bullets in a way that did not interfere with the player or cause the bullets to cease existing immediately. Other solutions may be preferable.